requirements.txt
aiohttp==3.9.3
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.32.49
backoff==2.2.1
boto3==1.34.49
botocore==1.34.49
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
continuous-threading==2.0.6
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.4
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
firebase-admin==6.4.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
GPUtil==1.4.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
humanfriendly==10.0
idna==3.6
imageio==2.34.0
inference==0.9.14
inference-cli==0.9.14
inference-sdk==0.9.14
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
networkx==3.2.1
numpy==1.25.2
onnxruntime==1.15.1
openai==1.12.0
opencv-python==4.8.0.76
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
pendulum==3.0.0
piexif==1.1.3
pillow==10.2.0
pluggy==1.4.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
PuLP==2.8.0
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.2
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyodbc==5.1.0
pyparsing==3.1.1
pyreadline3==3.4.1
pytest==8.0.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.0
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.1
six==1.16.0
skypilot==0.4.1
sniffio==1.3.1
starlette==0.36.3
supervision==0.18.0
sympy==1.12
tabulate==0.9.0
tifffile==2024.2.12
time-machine==2.13.0
tomli==2.0.1
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==1.7.0
xlrd==2.0.1
yarl==1.9.4
